Reorganize server modules
This should make it easier to add new routes in separate files.
