+++ tests/test_new_election.py
@@ -3,7 +3,8 @@ import pytest
 from flask.testing import FlaskClient
 from typing import Any
 
-from arlo_server import app, db, create_organization, UserType
+from arlo_server import app, db
+from arlo_server.routes import create_organization, UserType
 from tests.helpers import create_org_and_admin, set_logged_in_user, post_json
